feat: Add Todo update and delete functionality
This commit adds the necessary code changes to implement the update and delete functionality for Todo items. It includes the addition of `TodoUpdateView` and `TodoDeleteView` in the `todos/views.py` file, as well as corresponding URL patterns in the `setup/urls.py` file. The templates for the Todo form and Todo list have also been updated to include the appropriate links for editing and deleting Todo items.