feat: Add Todo complete functionality
This commit adds the necessary code changes to implement the complete functionality for Todo items. It includes the addition of `TodoCompleteView` in the `todos/views.py` file, as well as a corresponding URL pattern in the `setup/urls.py` file. The template for the Todo list has also been updated to include the appropriate link for marking a Todo item as done.